+++ pac4j-oauth/src/main/java/org/pac4j/oauth/client/GitHubClient.java
@@ -35,6 +35,7 @@ public class GitHubClient extends OAuth20Client {
     protected void clientInit() {
         configuration.setApi(GitHubApi.instance());
         configuration.setProfileDefinition(new GitHubProfileDefinition());
+        configuration.setTokenAsHeader(true);
         defaultLogoutActionBuilder((ctx, profile, targetUrl) ->
             Optional.of(RedirectionActionHelper.buildRedirectUrlAction(ctx, "https://github.com/logout")));