+++ documentation/docs/release-notes.md
@@ -6,6 +6,7 @@ title: Release notes&#58;
 **v4.0.2**:
 
 - Fix the `ClassCastException` when receiving a SAML logout response
+- Send the access token as a header for the `GithubClient`
 
 **v4.0.1**: